Added models, error handling.
