+++ pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "safepull"
 version = "1.1.0"
-description = "A quick and dirty command-line script to pull down and extract .py files out of tarballs and wheels without interfacing with the setup.py file."
+description = "A CLI tool for downloading and extracting packages from PyPI without interfacing with setup.py"
 readme = "README.md"
 license = { text = "MIT" }
 requires-python = ">=3.10"
